--- Compression/Compression.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.902
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Compression", "Compression\Compression.csproj", "{B1F2CDDE-4D92-4C7C-9FE5-4C43AB1D617C}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B1F2CDDE-4D92-4C7C-9FE5-4C43AB1D617C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B1F2CDDE-4D92-4C7C-9FE5-4C43AB1D617C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B1F2CDDE-4D92-4C7C-9FE5-4C43AB1D617C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B1F2CDDE-4D92-4C7C-9FE5-4C43AB1D617C}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {7A7B2CD6-9461-475A-849E-7EB138224431}
-	EndGlobalSection
-EndGlobal
