+++ file-compressor.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.28307.271
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FileSender", "src\FileSender\FileSender.csproj", "{B1F2CDDE-4D92-4C7C-9FE5-4C43AB1D617C}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{B1F2CDDE-4D92-4C7C-9FE5-4C43AB1D617C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B1F2CDDE-4D92-4C7C-9FE5-4C43AB1D617C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B1F2CDDE-4D92-4C7C-9FE5-4C43AB1D617C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B1F2CDDE-4D92-4C7C-9FE5-4C43AB1D617C}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {9A7A8E1A-A918-4D41-BE48-DBF1030A2909}
+	EndGlobalSection
+EndGlobal
